% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_language.R
\name{select_languageInput}
\alias{select_languageInput}
\title{Select Lanugage}
\usage{
select_languageInput(
  inputId,
  default_language,
  languages,
  icons_path,
  main_icon_path
)
}
\description{
Select Lanugage
}
