% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_language.R
\name{updateSelect_languageInput}
\alias{updateSelect_languageInput}
\title{Update Select Lanugage}
\usage{
updateSelect_languageInput(session, inputId, value, configuration = NULL)
}
\description{
Update Select Lanugage
}
